Everything you need to run the model
the files I was trying to push were too big and as a result I gathered 10 commits that couldn't push. I have to upload them as files from online github. Finished the model, trained the model, ran the model, tested the accuracy of the model, saved the training into a pickled file and wrote code to unpickle it in VS Code. I couldn't push for many days so this is a total of 15 hrs. The model is ready to use by commanding data.predict and inputing an image.